# backend/docs/contracts/generated-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with credentials",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"]
      }
    },
    "/auth/dev-login": {
      "post": {
        "operationId": "AuthController_devLogin",
        "summary": "Development login without OTP (only for development)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"]
      }
    },
    "/auth/keycloak-login": {
      "post": {
        "operationId": "AuthController_keycloakLogin",
        "summary": "Keycloak direct login (for Keycloak authentication)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOTP",
        "summary": "Verify OTP code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"]
      }
    },
...
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["IAM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/set-pin": {
      "post": {
        "operationId": "AuthController_setPin",
        "summary": "Set or replace the local six-digit PIN",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/conversations": {
      "get": {
        "operationId": "CommunicationController_list",
        "summary": "List user conversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CommunicationController_create",
        "summary": "Create a new conversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/conversations/{id}/messages": {
      "get": {
        "operationId": "CommunicationController_messages",
        "summary": "Get conversation messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CommunicationController_send",
        "summary": "Send a message to conversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/conversations/{id}/read": {
      "post": {
        "operationId": "CommunicationController_read",
        "summary": "Mark conversation as read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/conversations/stats/connections": {
      "get": {
        "operationId": "CommunicationController_getConnectionStats",
        "summary": "Get WebSocket connection statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Communication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/analytics/dashboard/overview": {
      "get": {
        "operationId": "AnalyticsDashboardController_getOverview",
        "summary": "Get dashboard overview stats",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Analytics Dashboard"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/analytics/dashboard/user-activity": {
      "get": {
        "operationId": "AnalyticsDashboardController_getUserActivity",
        "summary": "Get user activity trends",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Analytics Dashboard"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/analytics/dashboard/content-performance": {
      "get": {
        "operationId": "AnalyticsDashboardController_getContentPerformance",
        "summary": "Get content performance metrics",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Analytics Dashboard"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IRIB DWP API",
    "description": "IRIB East Azerbaijan Digital Workplace Platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "personnelCode": {
            "type": "string",
            "description": "Personnel code for registration",
            "example": "12345"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile phone number for registration",
            "example": "09123456789"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          }
        },
        "required": ["personnelCode", "mobile"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "personnelCode": {
            "type": "string",
            "description": "Personnel code for login",
            "example": "12345"
          },
          "password": {
            "type": "string",
            "description": "Password for login",
            "example": "password123"
          }
        },
        "required": ["personnelCode", "password"]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "description": "Challenge ID for OTP verification",
            "example": "abc123xyz"
          },
          "code": {
            "type": "string",
            "description": "6-digit OTP code",
            "example": "123456"
          }
        },
        "required": ["challengeId", "code"]
      },
      "SetPinDto": {
        "type": "object",
        "properties": {
          "pin": {
            "type": "string",
            "description": "6-digit PIN code",
            "example": "123456"
          }
        },
        "required": ["pin"]
      },
...
      "CreateConversationDto": {
        "type": "object",
        "properties": {}
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}